staticmethod fix for python < 3.10
+ factor our console.py out of cheatcodes.py
+ catch and log exceptions in stringify and console.log
